+++ abides-core/abides_core/abides.py
@@ -59,6 +59,7 @@ def run(
             "default_computation_delay",
             "custom_properties",
             "per_agent_computation_delays",
+            "skip_log",
         ],
     )